Support in-place loopback rewrites - WIP
