+++ pyron-core/src/main/scala/com/cloudentity/pyron/domain/http/UriPath.scala
@@ -1,0 +1,4 @@
+package com.cloudentity.pyron.domain.http
+
+case class UriPath(value: String) extends AnyVal
+
