+++ pyron-core/src/main/scala/com/cloudentity/pyron/rule/AppliedRewrite.scala
@@ -6,6 +6,7 @@ case class AppliedRewrite(path: String,
                           targetPath: String,
                           pathParams: PathParams,
                           from: PreparedRewrite) {
+
   def pathPrefix: String = from.pathPrefix
 
   def checkedPattern: String = from.checkedPattern
